Updated LETU logo.
Replaced the LETU logo with the LeTour Guide icon, & made sure that the connected users on the listening page works. The only bug is that the number of users only updates when the user is in an active channel, so if they enter an empty channel, it will still display the number of users from the last channel they were in. Otherwise, this feature works perfectly.